requirements_auto.txt
# Requirements automáticos para Sistema de Boletas
# Python 3.12

opencv-python==4.8.1.78
pytesseract==0.3.10
pdf2image==1.16.3
pillow==10.1.0
pandas==2.0.3
openpyxl==3.1.2
xlsxwriter==3.1.9
pypdf==3.17.0
numpy==1.24.3
